Add larger uints

+++ contracts/ContractDeployer.sol
@@ -13,7 +13,7 @@ contract ContractDeployer {
 
     mapping (address => address) public getLatestContractAddressOfOwner;
     mapping (address => address) public getContractOwner;
-    mapping (address => uint8) private _deployedContractsPerWallet;
+    mapping (address => uint256) private _deployedContractsPerWallet;
     mapping (address => address[]) private _trackingContracts;
 
     struct Addresses {
@@ -29,16 +29,16 @@ contract ContractDeployer {
      * @notice Function to deploy user's contracts
      * @param contractsAmount Amount of desired contracts to be deployed, which can not be 0 and can not exceed 5
      */
-    function deployContract(uint8 contractsAmount) external payable {
-        uint8 maxContractsPerTx = 5;
-        uint8 maxContractsPerWallet = 50;
-        uint56 deploymentPrice = 0.005 ether;
+    function deployContract(uint256 contractsAmount) external payable {
+        uint256 maxContractsPerTx = 5;
+        uint256 maxContractsPerWallet = 50;
+        uint256 deploymentPrice = 0.005 ether;
         require(tx.origin == msg.sender, "Only EOA");
         require(contractsAmount != 0 && contractsAmount <= maxContractsPerTx, "Max 5 contracts per tx");
         require(msg.value >= contractsAmount * deploymentPrice, "At least 0.005 ETH per contract");
         require(_deployedContractsPerWallet[msg.sender] + contractsAmount <= maxContractsPerWallet, "Max 50 contracts per EOA");
         _deployedContractsPerWallet[msg.sender] += contractsAmount;
-        for (uint8 i; i < contractsAmount;) {
+        for (uint256 i; i < contractsAmount;) {
             Contract deployedContract = new Contract();
             getContractWithOwnerAddress.push(Addresses({contractOwner: msg.sender, contractAddress: address(deployedContract)}));
             _trackingContracts[msg.sender] = contracts;
